좋아요한 게시물 조회 가능

+++ src/main/java/penterest/spring/domain/Like/service/LikeService.java
@@ -1,6 +1,8 @@
 package penterest.spring.domain.Like.service;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import penterest.spring.domain.Like.dto.LikeRequestDto;
@@ -11,8 +13,6 @@ import penterest.spring.domain.gif.repository.GifRepository;
 import penterest.spring.domain.member.entity.Member;
 import penterest.spring.domain.member.repository.MemberRepository;
 
-import java.util.Optional;
-
 @Service
 @RequiredArgsConstructor
 public class LikeService {
@@ -21,37 +21,54 @@ public class LikeService {
     private final MemberRepository memberRepository;
     private final GifRepository gifRepository;
 
-    @Transactional
-    public void like(LikeRequestDto likeRequestDto) throws Exception {
+    private String checkAuthority(Gif gif) {
+        //String loginUserEmail = SecurityUtil.getLoginUserEmail();
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String email = ((UserDetails) principal).getUsername();
 
-        Member member = memberRepository.findById(likeRequestDto.getMemberId())
-                .orElseThrow(()-> new Exception("Could not found member id : " + likeRequestDto.getMemberId()));
-
-        Gif gif = gifRepository.findById(likeRequestDto.getGifId())
-                .orElseThrow(()-> new Exception("Could not found member id : " + likeRequestDto.getGifId()));
-
-        Like existingLike = likeRepository.findByMemberAndGif(member, gif);
-        if (existingLike != null) {
-            throw new Exception("Like already exists.");
-        }
-
-        Like like = Like.builder()
-                .gif(gif)
-                .member(member)
-                .build();
-
-        likeRepository.save(like);
+        String writer = gif.getWriter().getEmail();
+        return writer;
     }
 
     @Transactional
-    public void dislike(LikeRequestDto likeRequestDto) throws Exception {
+    public String like(LikeRequestDto likeRequestDto) throws Exception {
+
         Member member = memberRepository.findById(likeRequestDto.getMemberId())
-                .orElseThrow(() -> new Exception("Could not found member id : " + likeRequestDto.getMemberId()));
-
+                .orElseThrow(() -> new Exception(("Could not found member id : " + likeRequestDto.getMemberId())));
         Gif gif = gifRepository.findById(likeRequestDto.getGifId())
-                .orElseThrow(() -> new Exception("Could not found member id : " + likeRequestDto.getGifId()));
+                .orElseThrow(()-> new Exception("Could not found gif id : " + likeRequestDto.getGifId()));
 
-        Like like = likeRepository.findByMemberAndGif(member, gif);
-        likeRepository.delete(like);
+        if (member.getEmail() == checkAuthority(gif)){
+            Like existingLike = likeRepository.findByMemberAndGif(member, gif);
+
+            if (existingLike != null) {
+                throw new Exception("Like already exists.");
+            }
+            Like like = Like.builder()
+                    .gif(gif)
+                    .member(member)
+                    .build();
+
+            likeRepository.save(like);
+        }else
+            throw new Exception("인가된 유저가 아닙니다");
+
+        return member.getEmail() + " 유저가 " + gif.getId() + " 게시물을 좋아합니다";
+    }
+
+    @Transactional
+    public String dislike(LikeRequestDto likeRequestDto) throws Exception {
+        Member member = memberRepository.findById(likeRequestDto.getMemberId())
+                .orElseThrow(() -> new Exception(("Could not found member id : " + likeRequestDto.getMemberId())));
+        Gif gif = gifRepository.findById(likeRequestDto.getGifId())
+                .orElseThrow(() -> new Exception("Could not found gif id : " + likeRequestDto.getGifId()));
+
+        if (member.getEmail() == checkAuthority(gif)) {
+            Like like = likeRepository.findByMemberAndGif(member, gif);
+            likeRepository.delete(like);
+        }else {
+            throw new Exception("인가된 유저가 아닙니다");
+        }
+        return member.getEmail() + " 유저가 " + gif.getId() + " 게시물을 좋아요를 취소했습니다";
     }
 }
